+++ infrastructure/src/main/java/io/devarium/infrastructure/persistence/repository/PostJpaRepository.java
@@ -1,6 +1,6 @@
-package io.devarium.infrastructure.repository;
+package io.devarium.infrastructure.persistence.repository;
 
-import io.devarium.infrastructure.entity.PostEntity;
+import io.devarium.infrastructure.persistence.entity.PostEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PostJpaRepository extends JpaRepository<PostEntity, Long> {
